requirements.txt
# Core dependencies for AI drug discovery pipeline
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Chemistry and molecular processing
rdkit>=2023.3.1
openbabel-wheel>=3.1.1
biopython>=1.79

# Machine Learning and Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.20.0
lightning>=2.0.0

# Docking and structural analysis
vina>=1.2.3
mdanalysis>=2.4.0
biotite>=0.37.0

# Web APIs and data collection
requests>=2.28.0
chembl-webresource-client>=0.10.0
pubchempy>=1.0.4

# Visualization
plotly>=5.10.0
pymol-open-source>=2.5.0
nglview>=3.0.3

# LLM integration
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.7.2

# Jupyter and development
jupyter>=1.0.0
notebook>=6.4.0
jupyterlab>=3.4.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=0.20.0

# Molecular dataset generation & fine-tuning
rdkit-pypi>=2023.3.1      # Chemistry toolkit for SMILES processing
torch-geometric>=2.3.0    # Graph neural networks
Pillow>=9.0.0             # Image processing for 2D molecular images
torchvision>=0.15.0       # Vision models and transforms
transformers>=4.20.0      # Foundation models (CLIP, ViT, etc.)

# Optional dependencies for advanced features
# deepchem>=2.7.1  # Uncomment for additional chemistry ML tools
# prody>=2.4.0     # Uncomment for protein dynamics analysis